Add files via upload
The website link is the index.html file, There are two buttons Ready to work and learn more, When clicked the ready to work will navigate to Log in/SignUp page, and the learn more button would navigate to additional information about FSM. I tried to make these buttons work but due to limited time, I could not finish. I've added transitions to these buttons as well as the images at the end of the website.